Move a bunch of core modules into a core sub-module
For better encapsulation.
