+++ src/core/device.rs
@@ -10,8 +10,9 @@ use std::{fmt, io};
 use libc::{dev_t, major, makedev, minor};
 use nix::sys::stat::SFlag;
 
-use crate::errors::ErrorKind;
 use crate::result::{DmError, DmResult};
+
+use crate::core::errors::ErrorKind;
 
 /// A struct containing the device's major and minor numbers
 ///
