+++ src/core/deviceinfo.rs
@@ -3,11 +3,10 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 use std::str::from_utf8;
 
-use crate::device::Device;
-use crate::dm_flags::DmFlags;
-use crate::dm_ioctl as dmi;
-use crate::types::{DmName, DmUuid};
-use crate::util::slice_to_null;
+use crate::core::{Device, DmFlags, DmName, DmUuid};
+
+use crate::core::dm_ioctl as dmi;
+use crate::core::util::slice_to_null;
 
 /// Name max length
 pub const DM_NAME_LEN: usize = 128;
