+++ src/core/dm_flags.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::dm_ioctl as dmi;
+use crate::core::dm_ioctl as dmi;
 
 bitflags! {
     /// Flags used by devicemapper.
