+++ src/core/dm_ioctl.rs
@@ -10,8 +10,9 @@
 use std::borrow::Cow;
 use std::fmt;
 
-use crate::device::Device;
-use crate::util::slice_to_null;
+use crate::core::Device;
+
+use crate::core::util::slice_to_null;
 
 pub type __s8 = ::libc::c_char;
 pub type __u8 = ::libc::c_uchar;
