+++ src/thinpooldev.rs
@@ -7,10 +7,10 @@ use std::fmt;
 use std::path::PathBuf;
 use std::str::FromStr;
 
-use crate::device::Device;
-use crate::deviceinfo::DeviceInfo;
-use crate::dm::DM;
-use crate::dm_options::DmOptions;
+use crate::core::{
+    DataBlocks, DevId, Device, DeviceInfo, DmName, DmOptions, DmUuid, MetaBlocks, Sectors,
+    TargetTypeBuf, DM,
+};
 use crate::lineardev::{LinearDev, LinearDevTargetParams};
 use crate::result::{DmError, DmResult, ErrorEnum};
 use crate::shared::{
@@ -18,10 +18,9 @@ use crate::shared::{
     make_unexpected_value_error, parse_device, parse_value, DmDevice, TargetLine, TargetParams,
     TargetTable,
 };
-use crate::types::{DataBlocks, DevId, DmName, DmUuid, MetaBlocks, Sectors, TargetTypeBuf};
 
 #[cfg(test)]
-use crate::device::devnode_to_devno;
+use crate::core::devnode_to_devno;
 #[cfg(test)]
 use std::path::Path;
 
@@ -683,8 +682,9 @@ pub fn minimal_thinpool(dm: &DM, path: &Path) -> ThinPoolDev {
 mod tests {
     use std::path::Path;
 
-    use crate::errors::{Error, ErrorKind};
     use crate::loopbacked::test_with_spec;
+
+    use crate::core::errors::{Error, ErrorKind};
 
     use super::*;
